SimCT/pyproject.toml
[tool.poetry]
name = "simct"
version = "0.1.0"
description = "Clinical study simulation platform"
authors = ["SimCT Team <team@example.com>"]
readme = "README.md"
packages = [{ include = "simct" }]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.12.5"
pyyaml = "^6.0.3"
pandas = "^3.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
ruff = "^0.15.5"
mypy = "^1.19.1"
ipython = "^9.11.0"
pandas-stubs = "^3.0.0.260204"
types-pyyaml = "^6.0.12.20250915"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.mypy]
python_version = "3.12"
strict = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
